Separate Landing into component-view
